feat(web-builder): extract the function `createPrimaryBuilder` and export `createStubBuilder`
