+++ packages/builder/web-builder/src/core/build.ts
@@ -1,8 +1,7 @@
 import { info, error, warn, formatWebpackStats } from '../shared';
-import { initConfigs, InitConfigsOptions } from './initConfigs';
 import type { WebpackConfig } from '../types';
 
-const webpackBuild = async (webpackConfigs: WebpackConfig[]) => {
+export const webpackBuild = async (webpackConfigs: WebpackConfig[]) => {
   const { default: webpack } = await import('webpack');
   const compiler = webpack(webpackConfigs);
 
@@ -36,14 +35,3 @@ const webpackBuild = async (webpackConfigs: WebpackConfig[]) => {
     });
   });
 };
-
-export async function build(options: InitConfigsOptions) {
-  const { context } = options;
-  const { webpackConfigs } = await initConfigs(options);
-
-  await context.hooks.onBeforeBuildHook.call({
-    webpackConfigs,
-  });
-  await webpackBuild(webpackConfigs);
-  await context.hooks.onAfterBuildHook.call();
-}
