+++ packages/builder/web-builder/src/shared/stub/context.ts
@@ -1,4 +1,4 @@
-import { createPrimaryContext } from '../../src/core/createContext';
+import { createPrimaryContext } from '../../core/createContext';
 import type { StubBuilderOptions } from './builder';
 
 export function createStubContext(options: Required<StubBuilderOptions>) {
