+++ packages/builder/web-builder/tests/plugins/bundleAnalyzer.test.ts
@@ -1,6 +1,6 @@
 import { expect, describe, it } from 'vitest';
 import { PluginBundleAnalyzer } from '../../src/plugins/bundleAnalyzer';
-import { createStubBuilder } from '../utils/builder';
+import { createStubBuilder } from '../../src/shared/stub';
 
 describe('plugins/bundleAnalyze', () => {
   it('should add bundle analyze plugin', async () => {
