+++ packages/builder/web-builder/tests/plugins/manifest.test.ts
@@ -1,15 +1,16 @@
 import { expect, describe, it } from 'vitest';
 import { PluginManifest } from '../../src/plugins/manifest';
-import { createStubBuilder } from '../utils/builder';
-import { isPluginRegistered } from '../utils/webpack';
+import { createStubBuilder } from '../../src/shared/stub';
 
 describe('plugins/manifest', () => {
   it('should not register manifest plugin by default', async () => {
     const builder = createStubBuilder({
       plugins: [PluginManifest()],
     });
-    const config = await builder.unwrapWebpackConfig();
-    expect(isPluginRegistered(config, 'WebpackManifestPlugin')).toBeFalsy();
+
+    expect(
+      await builder.matchWebpackPlugin('WebpackManifestPlugin'),
+    ).toBeFalsy();
   });
 
   it('should register manifest plugin when output.enableAssetManifest is enabled', async () => {
@@ -22,7 +23,8 @@ describe('plugins/manifest', () => {
       },
     });
 
-    const config = await builder.unwrapWebpackConfig();
-    expect(isPluginRegistered(config, 'WebpackManifestPlugin')).toBeTruthy();
+    expect(
+      await builder.matchWebpackPlugin('WebpackManifestPlugin'),
+    ).toBeTruthy();
   });
 });
