+++ packages/builder/web-builder/tests/plugins/react.test.ts
@@ -1,12 +1,10 @@
 import { expect, describe, it } from 'vitest';
-import { setPathSerializer } from '../utils/snapshot';
 import { PluginReact } from '../../src/plugins/react';
 import { PluginBabel } from '../../src/plugins/babel';
 import { PluginTsLoader } from '../../src/plugins/tsLoader';
-import { createStubBuilder } from '../utils/builder';
+import { createStubBuilder } from '../../src/shared/stub';
 
 describe('plugins/react', () => {
-  setPathSerializer();
   it('should work with babel-loader', async () => {
     const builder = createStubBuilder({
       plugins: [PluginBabel(), PluginReact()],
