+++ packages/builder/web-builder/tests/plugins/svg.test.ts
@@ -1,10 +1,8 @@
 import { expect, describe, it } from 'vitest';
-import { setPathSerializer } from '../utils/snapshot';
 import { PluginSvg } from '../../src/plugins/svg';
-import { createStubBuilder } from '../utils/builder';
+import { createStubBuilder } from '../../src/shared/stub';
 
 describe('plugins/svg', () => {
-  setPathSerializer();
   it('export default url', async () => {
     const builder = createStubBuilder({
       plugins: [PluginSvg()],
