+++ packages/builder/web-builder/tests/setup.ts
@@ -1,0 +1,23 @@
+import { normalizeToPosixPath } from '@modern-js/utils';
+import { expect } from 'vitest';
+
+expect.addSnapshotSerializer({
+  test: val => typeof val === 'string' && /\\/g.test(val),
+  print: val => `"${normalizeToPosixPath(val as string)}"`,
+});
+
+expect.addSnapshotSerializer({
+  test: val => typeof val === 'string' && val.includes('node_modules'),
+  print: val =>
+    `"<NODE_MODULES>${normalizeToPosixPath(
+      (val as string).split('node_modules').pop(),
+    )}"`,
+});
+
+expect.addSnapshotSerializer({
+  test: val => typeof val === 'string' && val.includes('packages'),
+  print: val =>
+    `"<ROOT>/packages${normalizeToPosixPath(
+      (val as string).split('packages').pop(),
+    )}"`,
+});
