+++ packages/builder/web-builder/tests/tsconfig.json
@@ -7,5 +7,5 @@
     "baseUrl": "./",
     "isolatedModules": true,
     "paths": {}
-  }
+  },
 }
