--- packages/builder/web-builder/tests/utils/matchLoader.ts
@@ -1,44 +1,0 @@
-import type { Configuration } from 'webpack';
-
-/**
- * The util function is used to match the loader in the webpack config.
- * @author yangxingyuan
- * @param {Configuration} config - The webpack config.
- * @param {string} loader - The name of loader.
- * @param {string}  testFile - The name of test file that will be handled by webpack.
- * @returns {boolean} The result of the match.
- */
-export function matchLoader({
-  config,
-  loader,
-  testFile,
-}: {
-  config: Configuration;
-  loader: string;
-  testFile: string;
-}): boolean {
-  if (!config.module?.rules) {
-    return false;
-  }
-  return config.module.rules.some(rule => {
-    if (
-      typeof rule === 'object' &&
-      rule.test &&
-      rule.test instanceof RegExp &&
-      rule.test.test(testFile)
-    ) {
-      return (
-        Array.isArray(rule.use) &&
-        rule.use.some(useOptions => {
-          if (typeof useOptions === 'object' && useOptions !== null) {
-            return useOptions.loader?.includes(loader);
-          } else if (typeof useOptions === 'string') {
-            return useOptions.includes(loader);
-          }
-          return false;
-        })
-      );
-    }
-    return false;
-  });
-}
