--- packages/builder/web-builder/tests/utils/webpack.ts
@@ -1,4 +1,0 @@
-import type { WebpackConfig } from '../../src/types';
-
-export const isPluginRegistered = (config: WebpackConfig, pluginName: string) =>
-  config.plugins?.some(item => item.constructor.name === pluginName);
