+++ packages/builder/web-builder/vitest.config.ts
@@ -5,6 +5,7 @@ const config = defineConfig({
   test: {
     root: __dirname,
     environment: 'node',
+    setupFiles: ['./tests/setup.ts'],
   },
 });
